pyproject.toml
[project]
name = "fermenter-temp-controller"
version = "1.0.0"
description = "Modernised (2025) version of Fermentation Temperature Controller"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "influxdb-client",
  "pydantic",
  "pyserial",
  "pyserial-asyncio",
  "pytz",
  "pyzmq",
  "requests",
  "validators",
]

[dependency-groups]
help = [
    "pytest>=8.3.5",
]
test = [
    "pytest-asyncio>=0.26.0",
]
